Initial SM gui work
git-svn-id: svn://scribus.net/branches/Version13x/Scribus@5850 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/pageitem_textframe.cpp
@@ -169,7 +169,7 @@ static void dumpIt(const ParagraphStyle& pstyle, QString indent = QString("->"))
 	qDebug(QString("%6%1/%2 @ %3: %4--%5")
 		   .arg(pstyle.name())
 		   .arg(pstyle.parent()? pstyle.parent()->name() : "-")
-		   .arg( (unsigned int) &pstyle)
+		   .arg( (unsigned long int) &pstyle)
 		   .arg(pstyle.leftMargin())
 		   .arg(pstyle.rightMargin())
 		   .arg(indent));
